LayoutTests/fast/text/line-separator-as-segment-break-expected.txt

U+2028 LINE SEPARATOR is a forced line break in all white-space modes (UAX#14 class BK, CSS Text Level 3 Â§5.1).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('normal').offsetHeight is 20
PASS document.getElementById('pre').offsetHeight is 20
PASS document.getElementById('pre-wrap').offsetHeight is 20
PASS document.getElementById('pre-line').offsetHeight is 20
PASS successfullyParsed is true

TEST COMPLETE
a b
a b
a b
a b
